+++ mica-caffeine/src/main/java/net/dreamlu/mica/caffeine/config/CaffeineAutoCacheManager.java
@@ -18,11 +18,12 @@ package net.dreamlu.mica.caffeine.config;
 
 import com.github.benmanes.caffeine.cache.CacheLoader;
 import com.github.benmanes.caffeine.cache.Caffeine;
-import net.dreamlu.mica.core.utils.ReflectUtil;
+import com.github.benmanes.caffeine.cache.CaffeineSpec;
 import net.dreamlu.mica.core.utils.StringPool;
 import org.springframework.boot.convert.DurationStyle;
 import org.springframework.cache.caffeine.CaffeineCacheManager;
 import org.springframework.lang.Nullable;
+import org.springframework.util.ReflectionUtils;
 
 import java.lang.reflect.Field;
 import java.time.Duration;
@@ -35,15 +36,15 @@ import java.util.Objects;
  * @author L.cm
  */
 public class CaffeineAutoCacheManager extends CaffeineCacheManager {
-	private static final Field CACHE_BUILDER_FIELD;
 	private static final Field CACHE_LOADER_FIELD;
 
 	static {
-		CACHE_BUILDER_FIELD = Objects.requireNonNull(ReflectUtil.getField(CaffeineCacheManager.class, "cacheBuilder"));
-		CACHE_BUILDER_FIELD.setAccessible(true);
-		CACHE_LOADER_FIELD = Objects.requireNonNull(ReflectUtil.getField(CaffeineCacheManager.class, "cacheLoader"));
+		CACHE_LOADER_FIELD = Objects.requireNonNull(ReflectionUtils.findField(CaffeineCacheManager.class, "cacheLoader"));
 		CACHE_LOADER_FIELD.setAccessible(true);
 	}
+
+	@Nullable
+	private CaffeineSpec caffeineSpec = null;
 
 	public CaffeineAutoCacheManager() {
 		super();
@@ -53,23 +54,27 @@ public class CaffeineAutoCacheManager extends CaffeineCacheManager {
 		super(cacheNames);
 	}
 
-	@SuppressWarnings("unchecked")
-	protected Caffeine<Object, Object> getCacheBuilder() {
-		try {
-			return (Caffeine<Object, Object>) CACHE_BUILDER_FIELD.get(this);
-		} catch (IllegalAccessException e) {
-			throw new IllegalArgumentException(e);
-		}
-	}
-
 	@Nullable
 	@SuppressWarnings("unchecked")
 	protected CacheLoader<Object, Object> getCacheLoader() {
-		try {
-			return (CacheLoader<Object, Object>) CACHE_LOADER_FIELD.get(this);
-		} catch (IllegalAccessException e) {
-			return null;
-		}
+		return (CacheLoader<Object, Object>) ReflectionUtils.getField(CACHE_LOADER_FIELD, this);
+	}
+
+	@Override
+	public void setCaffeine(Caffeine<Object, Object> caffeine) {
+		throw new IllegalArgumentException("mica-caffeine not support customization Caffeine bean，you can customize CaffeineSpec bean.");
+	}
+
+	@Override
+	public void setCaffeineSpec(CaffeineSpec caffeineSpec) {
+		super.setCaffeineSpec(caffeineSpec);
+		this.caffeineSpec = caffeineSpec;
+	}
+
+	@Override
+	public void setCacheSpecification(String cacheSpecification) {
+		super.setCacheSpecification(cacheSpecification);
+		this.caffeineSpec = CaffeineSpec.parse(cacheSpecification);
 	}
 
 	/**
@@ -88,7 +93,12 @@ public class CaffeineAutoCacheManager extends CaffeineCacheManager {
 		}
 		// 转换时间，支持时间单位例如：300ms，第二个参数是默认单位
 		Duration duration = DurationStyle.detectAndParse(cacheArray[1], ChronoUnit.SECONDS);
-		Caffeine<Object, Object> cacheBuilder = getCacheBuilder();
+		Caffeine<Object, Object> cacheBuilder;
+		if (this.caffeineSpec != null) {
+			cacheBuilder = Caffeine.from(caffeineSpec);
+		} else {
+			cacheBuilder = Caffeine.newBuilder();
+		}
 		CacheLoader<Object, Object> cacheLoader = getCacheLoader();
 		if (cacheLoader == null) {
 			return cacheBuilder.expireAfterAccess(duration).build();