+++ mica-core/src/main/java/net/dreamlu/mica/core/compiler/ByteCodeLoader.java
@@ -17,6 +17,7 @@
 package net.dreamlu.mica.core.compiler;
 
 import net.dreamlu.mica.core.function.CheckedFunction;
+import net.dreamlu.mica.core.utils.CollectionUtil;
 import net.dreamlu.mica.core.utils.Unchecked;
 
 import java.security.SecureClassLoader;
@@ -32,7 +33,7 @@ public class ByteCodeLoader extends SecureClassLoader {
 	/**
 	 * Map which represents class name and its compiled java object
 	 */
-	private static final ConcurrentMap<String, Class<?>> javaFileObjectMap = new ConcurrentHashMap<>();
+	private static final ConcurrentMap<String, Class<?>> JAVA_FILE_OBJECT_MAP = new ConcurrentHashMap<>();
 	private final String className;
 	private final byte[] byteCode;
 
@@ -67,7 +68,7 @@ public class ByteCodeLoader extends SecureClassLoader {
 	 */
 	public static Class<?> load(String className, byte[] byteCode) {
 		CheckedFunction<String, Class<?>> classLoadFunc = (key) -> new ByteCodeLoader(key, byteCode).loadClass(className);
-		return javaFileObjectMap.computeIfAbsent(className, Unchecked.function(classLoadFunc));
+		return CollectionUtil.computeIfAbsent(JAVA_FILE_OBJECT_MAP, className, Unchecked.function(classLoadFunc));
 	}
 
 	/**