+++ mica-logging/src/main/java/net/dreamlu/mica/logging/appender/LoggingLogStashAppender.java
@@ -28,7 +28,6 @@ import net.logstash.logback.encoder.LogstashEncoder;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.env.Environment;
 
-import java.net.InetSocketAddress;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -46,11 +45,11 @@ public class LoggingLogStashAppender implements ILoggingAppender {
 	public LoggingLogStashAppender(Environment environment,
 								   MicaLoggingProperties properties) {
 		this.properties = properties;
-		// 1. 服务名和环境和日志目录
+		// 1. 服务名和环境
 		String appName = environment.getRequiredProperty(MicaConstant.SPRING_APP_NAME_KEY);
 		String profile = environment.getRequiredProperty(MicaConstant.ACTIVE_PROFILES_PROPERTY);
 		// 2. json 自定义字段
-		Map<String, Object> customFields = new HashMap<>();
+		Map<String, Object> customFields = new HashMap<>(4);
 		customFields.put("appName", appName);
 		customFields.put("profile", profile);
 		this.customFieldsJson = JsonUtil.toJson(customFields);
@@ -84,12 +83,12 @@ public class LoggingLogStashAppender implements ILoggingAppender {
 	 * @param customFields       a {@link String} object.
 	 * @param logStashProperties a {@link net.dreamlu.mica.logging.config.MicaLoggingProperties.Logstash} object.
 	 */
-	public static void addLogStashTcpSocketAppender(LoggerContext context,
+	private static void addLogStashTcpSocketAppender(LoggerContext context,
 													String customFields,
 													MicaLoggingProperties.Logstash logStashProperties) {
 		// More documentation is available at: https://github.com/logstash/logstash-logback-encoder
 		final LogstashTcpSocketAppender logStashAppender = new LogstashTcpSocketAppender();
-		logStashAppender.addDestinations(new InetSocketAddress(logStashProperties.getHost(), logStashProperties.getPort()));
+		logStashAppender.addDestination(logStashProperties.getDestinations());
 		logStashAppender.setContext(context);
 		logStashAppender.setEncoder(logstashEncoder(customFields));
 		logStashAppender.setName(ASYNC_LOG_STASH_APPENDER_NAME);